Add files via upload
Used a face-classifier to detect faces, then trained another to detect different objects, depending on color. That goal was to be able to detect colorless objects that are also minute- colorless contains not 3 but 4 color channels, which makes it exponentially harder to detect this objects as the color channels extend beyond RGB.
face_detection.py and object_detection.py contains scripts that enable detection of objects from live video streams or still images.